test: route assert_select through Dom for kotlin/swift/crystal/go
Completes the Dom primitive contract across all nine targets (the
first five — ruby/python/rust/elixir/typescript — landed in the
earlier Dom commit). Same shape everywhere: assert_select queries a
small Dom surface (parse(html)->doc, select(root,selector)->[node],
text(node)->string) instead of inlining substring matching, with a
behavior-preserving stub (doc and node are both the html string;
select fabricates one synthetic node per fragment occurrence).

- kotlin / swift: add a Dom object/enum in their now-file-based
  test_support, route the four assertSelect overloads through it, drop
  the per-class selectorFragment. (Possible only because the prior
  commit extracted these runtimes to runtime/{kotlin,swift}/ files.)

- crystal: collapse the two duplicate test surfaces onto ONE shared
  Roundhouse::Dom (defined in test_support.cr). Both the live spec base
  (RoundhouseTest in test_helper.cr) and the otherwise-dead TestResponse
  in test_support.cr now route through it, eliminating the two copies of
  selector_fragment.

- go: promote the test-support runtime from an inline Rust string blob
  in src/emit/go2.rs to runtime/go/v2/test_support_test.go (matching the
  other ten Go runtime files, which are include_str!'d), then add a Dom
  stub (domParse/domSelect/domText) and route AssertSelect/Text/Min
  through it. Pure relocation + Dom rewire; emitted output unchanged
  except the Dom indirection.

Net: assert_select logic is now uniform across targets, and a real CSS
engine (jsoup / SwiftSoup / XML::Node / goquery / scraper / cheerio /
lxml / Floki / Nokogiri / lexbor-FFI) can be dropped into any single
target's Dom without touching call sites or other targets.

Verified: crystal/go/kotlin/swift toolchain tests all pass (the spec/
go-test runs exercise assert_select through the new Dom); default cargo
test green (61 ok / 0 failed). gofmt clean; crystal diff limited to the
assert_select rewire (reverted a local-formatter-version churn on two
unrelated lines).
