+++ src/emit/go2.rs
@@ -729,166 +729,7 @@ fn rewrite_test_file_to_v2(f: EmittedFile, model_names: &[String]) -> EmittedFil
 /// emit references. Dispatches through v2's `Dispatch` (so it
 /// exercises the same router + controllers production traffic uses).
 fn emit_v2_test_support(_app: &App) -> EmittedFile {
-    let content = r##"// Generated by Roundhouse (go2 Phase 6 step 1).
-// Test-support runtime for `go test ./app/v2/` — Dispatches through
-// the same v2 router/controller path production traffic uses. Mirrors
-// the legacy `app/test_support.go` API surface (TestClient.Get/Post/
-// Patch/Put/Delete, TestResponse.AssertOk/AssertSelect/...) so the
-// legacy spec.rs emit doesn't have to know it's targeting v2.
-
-package v2
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"strings"
-	"testing"
-)
-
-// (SetupTestDB lives in db.go — it opens a fresh :memory: SQLite
-// and runs schema DDL. fixtures_test.go's setupTest() calls it
-// before reloading fixtures.)
-
-// TestResponse — Rails-Minitest-shaped assertion helpers over a
-// response captured via httptest.ResponseRecorder.
-type TestResponse struct {
-	t        *testing.T
-	Body     string
-	Status   int
-	Location string
-}
-
-func (r *TestResponse) AssertOk() {
-	r.t.Helper()
-	if r.Status != 200 {
-		r.t.Fatalf("expected 200 OK, got %d", r.Status)
-	}
-}
-
-func (r *TestResponse) AssertUnprocessable() {
-	r.t.Helper()
-	if r.Status != 422 {
-		r.t.Fatalf("expected 422 Unprocessable Entity, got %d", r.Status)
-	}
-}
-
-func (r *TestResponse) AssertStatus(code int) {
-	r.t.Helper()
-	if r.Status != code {
-		r.t.Fatalf("expected status %d, got %d", code, r.Status)
-	}
-}
-
-func (r *TestResponse) AssertRedirectedTo(path string) {
-	r.t.Helper()
-	if r.Status < 300 || r.Status >= 400 {
-		r.t.Fatalf("expected a redirection, got %d", r.Status)
-	}
-	if !strings.Contains(r.Location, path) {
-		r.t.Fatalf("expected Location to contain %q, got %q", path, r.Location)
-	}
-}
-
-func (r *TestResponse) AssertSelect(selector string) {
-	r.t.Helper()
-	frag := selectorFragment(selector)
-	if !strings.Contains(r.Body, frag) {
-		r.t.Fatalf("expected body to match selector %q (looked for %q)", selector, frag)
-	}
-}
-
-func (r *TestResponse) AssertSelectText(selector, text string) {
-	r.t.Helper()
-	r.AssertSelect(selector)
-	if !strings.Contains(r.Body, text) {
-		r.t.Fatalf("expected body to contain text %q under selector %q", text, selector)
-	}
-}
-
-func (r *TestResponse) AssertSelectMin(selector string, n int) {
-	r.t.Helper()
-	frag := selectorFragment(selector)
-	count := strings.Count(r.Body, frag)
-	if count < n {
-		r.t.Fatalf("expected at least %d matches for selector %q, got %d", n, selector, count)
-	}
-}
-
-func selectorFragment(selector string) string {
-	first := strings.Fields(selector)
-	if len(first) == 0 {
-		return ""
-	}
-	head := first[0]
-	switch {
-	case strings.HasPrefix(head, "#"):
-		return "id=\"" + head[1:] + "\""
-	case strings.HasPrefix(head, "."):
-		return head[1:] + "\""
-	default:
-		return "<" + head
-	}
-}
-
-// TestClient — in-process HTTP client routed through Router() (the
-// same handler production main.go boots). Uses httptest's
-// ResponseRecorder so we don't open a real listener.
-type TestClient struct {
-	t *testing.T
-}
-
-func NewTestClient(t *testing.T) *TestClient {
-	return &TestClient{t: t}
-}
-
-func (c *TestClient) Get(path string) *TestResponse {
-	return c.dispatch("GET", path, nil)
-}
-
-func (c *TestClient) Post(path string, body map[string]string) *TestResponse {
-	return c.dispatch("POST", path, body)
-}
-
-func (c *TestClient) Patch(path string, body map[string]string) *TestResponse {
-	return c.dispatch("PATCH", path, body)
-}
-
-func (c *TestClient) Put(path string, body map[string]string) *TestResponse {
-	return c.dispatch("PUT", path, body)
-}
-
-func (c *TestClient) Delete(path string) *TestResponse {
-	return c.dispatch("DELETE", path, nil)
-}
-
-func (c *TestClient) dispatch(method, path string, body map[string]string) *TestResponse {
-	c.t.Helper()
-	var bodyReader *strings.Reader
-	if body != nil {
-		form := url.Values{}
-		for k, v := range body {
-			form.Set(k, v)
-		}
-		bodyReader = strings.NewReader(form.Encode())
-	}
-	var req *http.Request
-	if bodyReader != nil {
-		req = httptest.NewRequest(method, path, bodyReader)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	} else {
-		req = httptest.NewRequest(method, path, nil)
-	}
-	w := httptest.NewRecorder()
-	Router().ServeHTTP(w, req)
-	return &TestResponse{
-		t:        c.t,
-		Body:     w.Body.String(),
-		Status:   w.Code,
-		Location: w.Header().Get("Location"),
-	}
-}
-"##;
+    let content = include_str!("../../runtime/go/v2/test_support_test.go");
     EmittedFile {
         // `_test.go` suffix keeps this out of `go build` — the
         // overrides + helpers only exist during `go test` and don't
